Complete Authorization

+++ backend/controllers/goalControllers.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel')
 const { set } = require('mongoose');
 
 //@desc Get Goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler( async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user._id });
     res.status(200).json(goals);
 })
 
@@ -22,7 +23,8 @@ const setGoal = asyncHandler( async (req, res) => {
         if(goal) res.status(200).json({message: 'This goal is already included!'});
         else {
             const goal = new Goal({
-                text: req.body.text
+                text: req.body.text,
+                user: req.user._id
             });
             await goal.save();
             res.status(200).json(goal);
@@ -40,7 +42,20 @@ const updateGoal = asyncHandler( async (req, res) => {
         throw new Error('Goal Not Found!');
     }
     else {
-       const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,
+        const user = User.findById(req.user._id);
+
+        //Check for user
+        if(!user) {
+            res.status(401)
+            throw new Error('User not found')
+        }
+        //Make sure that logged in user matches the goal user
+        if(goalToUpdate.user.toString() !== user._id) {
+            res.status(401)
+            throw new Error('User is not authorized!')
+        }
+
+        const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,
          {
             new: true
         });
@@ -53,10 +68,24 @@ const updateGoal = asyncHandler( async (req, res) => {
 //@access Private
 const deleteGoal = asyncHandler( async (req, res) => {
     const goalToDelete = await Goal.findOne({_id: req.params.id});
+    const user = await User.findById({_id: req.user._id});
+
     if(!goalToDelete) {
         res.status(400)
         throw new Error('Goal Not Found!');
     }
+
+     //Check for user
+     if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure that logged in user matches the goal user
+    if(goalToDelete.user.toString() != req.user._id) {
+        res.status(401)
+        throw new Error('User is not authorized!')
+    }
+
     const deletedGoal = await Goal.findByIdAndRemove(goalToDelete._id);
     res.status(200).json({id: deletedGoal.id});
 })
